+++ src/database-pgsql/src/PostgreSqlSwooleExtConnection.php
@@ -28,11 +28,6 @@ class PostgreSqlSwooleExtConnection extends Connection
 {
     use PostgreSqlSwooleExtManagesTransactions;
 
-    /**
-     * @var PostgreSQL
-     */
-    protected mixed $pdo;
-
     protected int $fetchMode = SW_PGSQL_ASSOC;
 
     /**
@@ -57,8 +52,6 @@ class PostgreSqlSwooleExtConnection extends Connection
                 return true;
             }
 
-            $this->getPdo();
-
             $statement = $this->prepare($query);
 
             $result = $statement->execute($this->prepareBindings($bindings));
@@ -81,8 +74,6 @@ class PostgreSqlSwooleExtConnection extends Connection
             if ($this->pretending()) {
                 return 0;
             }
-
-            $this->getPdo();
 
             $statement = $this->prepare($query);
             $this->recordsHaveBeenModified();
@@ -110,14 +101,12 @@ class PostgreSqlSwooleExtConnection extends Connection
                 return [];
             }
 
-            $this->getPdoForSelect($useReadPdo);
-
-            $statement = $this->prepare($query);
+            $statement = $this->prepare($query, $useReadPdo);
 
             $result = $statement->execute($this->prepareBindings($bindings));
 
-            if ($result === false || ! empty($this->pdo->error)) {
-                throw new QueryException($query, [], new Exception($this->pdo->error));
+            if ($result === false || ! empty($statement->error)) {
+                throw new QueryException($query, [], new Exception($statement->error));
             }
 
             return $statement->fetchAll($this->fetchMode) ?: [];
@@ -134,9 +123,7 @@ class PostgreSqlSwooleExtConnection extends Connection
                 return [];
             }
 
-            $this->getPdoForSelect($useReadPdo);
-
-            $statement = $this->prepare($query);
+            $statement = $this->prepare($query, $useReadPdo);
 
             $statement->execute($this->prepareBindings($bindings));
 
@@ -161,7 +148,7 @@ class PostgreSqlSwooleExtConnection extends Connection
 
         $result = $statement->execute($bindings);
         if (! $result) {
-            throw new QueryException($query, [], new Exception($this->pdo->error));
+            throw new QueryException($query, [], new Exception($statement->error));
         }
 
         return $statement->fetchAll(SW_PGSQL_ASSOC);
@@ -228,16 +215,18 @@ class PostgreSqlSwooleExtConnection extends Connection
         return new PostgresDriver();
     }
 
-    protected function prepare(string $query): PostgreSQLStatement
+    protected function prepare(string $query, bool $useReadPdo = true): PostgreSQLStatement
     {
         $num = 1;
         while (strpos($query, '?')) {
             $query = $this->str_replace_once('?', '$' . $num++, $query);
         }
 
-        $statement = $this->pdo->prepare($query);
+        /** @var PostgreSQL $pdo */
+        $pdo = $this->getPdoForSelect($useReadPdo);
+        $statement = $pdo->prepare($query);
         if (! $statement) {
-            throw new QueryException($query, [], new Exception($this->pdo->error));
+            throw new QueryException($query, [], new Exception($pdo->error));
         }
 
         return $statement;